Agregado el metodo GET de ArchivosMultimediaController y correccion de tipos

+++ webapi/Data/ArchivosMultimediaDAO.cs
@@ -23,7 +23,7 @@ namespace webapi.Data
             }
         }
 
-        public ArchivosMultimedia GetOneById(ulong idArchivo)
+        public ArchivosMultimedia GetOneById(int idArchivo)
         {
             var ad = new AccesoDatos();
             ArchivosMultimedia archivoMultimedia = null;
